infoTiles/moved-folder
- moved folder to ./src/Components/Dashboard from (...)/Dashboard/Hero
